[#2437] Upgrade fixed-data-table to 1.0.1 and trying ColumnGroup
