+++ client/package.json
@@ -67,7 +67,7 @@
     "extract-text-webpack-plugin": "3.0.1",
     "file-loader": "0.11.2",
     "file-saver": "^1.3.8",
-    "fixed-data-table-2": "^0.7.17",
+    "fixed-data-table-2": "^1.0.1",
     "history": "^4.7.2",
     "html-webpack-plugin": "2.29.0",
     "http-proxy-middleware": "^0.17.4",
