+++ client/src/components/dataset/ColumnHeader.jsx
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import Immutable from 'immutable';
 import { FormattedMessage } from 'react-intl';
 
 require('./ColumnHeader.scss');
@@ -10,8 +9,6 @@ export default class ColumnHeader extends Component {
   constructor() {
     super();
     this.handleDataTypeMenuClick = this.handleDataTypeMenuClick.bind(this);
-    this.handleColumnMenuClick = this.handleColumnMenuClick.bind(this);
-    this.handleRemoveSort = this.handleRemoveSort.bind(this);
   }
 
   handleDataTypeMenuClick(event) {
@@ -33,61 +30,16 @@ export default class ColumnHeader extends Component {
     });
   }
 
-  handleColumnMenuClick() {
-    const el = this.columnHeaderContainer;
-    const rect = el.getBoundingClientRect();
-    const verticalOffset = rect.top + el.offsetHeight;
-    const horizontalOffset = rect.left;
-    const width = el.offsetWidth;
-
-    const dimensions = {
-      left: horizontalOffset,
-      top: verticalOffset,
-      width,
-    };
-    this.props.onToggleColumnContextMenu({
-      dimensions,
-      column: this.props.column,
-    });
-  }
-
-  handleRemoveSort(event, column) {
-    event.stopPropagation();
-    this.props.onRemoveSort(Immutable.fromJS({
-      op: 'core/remove-sort',
-      args: { columnName: column.get('columnName') },
-      onError: 'fail',
-    }));
-  }
 
   render() {
-    const { column, disabled } = this.props;
+    const { column } = this.props;
 
     return (
       <div
-        className={`ColumnHeader ${disabled ? '' : 'clickable'}
-          ${this.props.columnMenuActive ? 'columnMenuActive' : ''}`}
-        ref={(ref) => { this.columnHeaderContainer = ref; }}
-        onClick={this.handleColumnMenuClick}
+        className="ColumnHeader"
       >
-        {column.get('sort') != null ?
-          <div
-            className="sortLabel"
-          >
-            Sort: {column.get('direction') === 'ASC' ? 'Ascending' : 'Descending'}
-            <span
-              className="cancelSort"
-              onClick={event => this.handleRemoveSort(event, column)}
-            >
-              ✕
-            </span>
-          </div>
-          : null
-        }
         <span
           className="columnTitleText"
-          title={column.get('title')}
-          data-test-id={column.get('title')}
         >
           {column.get('key') ?
             <span className="columnKey">
@@ -105,7 +57,6 @@ export default class ColumnHeader extends Component {
               </span>
             </span>
           }
-          {column.get('title')}
         </span>
       </div>
     );
